V2.0.0.0
V2 introduces finer grained control of access to objects via group protection.

As part of this change the interfaces have changed and new interfaces for Groups have been added.

A distinction is now made between the group that a subject belongs to and the groups that a subject has. The subject groups define the group ID together with the application operation.

This allows a subject to be granted read access to an object via group protection whilst being denied write, delete or execute. The group permission on the object is still definitive; in that a group may be applicable to write - but this will not grant write access if the group protection on the object prevents writing.

The subject interface must implement IsGroupEquivalent(Operation operation, IControlledObject obj)

typically this would be implemented as follows.

            return obj.Groups.Where(og => Groups.Any(xx => xx.Id == og.Id))
                .Any(og => ReferenceMonitor.IsOperationPermitted(operation, og.ApplicableTo));

The reference monitor could implement this directly; however it is this way to allow any implicity group membership that may be required - such as when a subject is implicitly granted permissions because they are contained within another object, i.e. to access objects that may be owned by their parent using group protection.