Group operations.
Interface changes to allow a group to define the operations to which it applies.

Whilst writing the unit test it became clear that the base interface for IControlledObjectGroup also needs to define the operation(s) that the group can be applied to.

This builds on the previous changes.

Operation is implemented as a bitfield enum; the original reason for this dates back to the original version of the code - which was in C++ and as such list operations were more expensive.